backend/requirements.txt
django>=3.2.9,<4
pillow>=8.4.0
graphene-django>=2.15.0
django-filter>=21.1
str2bool>=1.1
dj-database-url>=0.5.0
whitenoise>=5.3.0
gunicorn>=20.1.0
mysqlclient>=2.0.3
django-ckeditor==6.3.0
django-extensions>=3.1.5
django-reversion-compare>=0.14.1
django-imagekit>=4.1.0
django-ordered-model>=3.4.3
raven>=6.10.0
django-allauth>=0.42.6
django-phonenumber-field>=6.0.0
phonenumberslite>=8.12.36
python-memcached>=1.59
django-import-export>=2.6.1
tablib[all]>=3.1.0
django-modeladmin-reorder>=0.3.1
django-private-storage>=2.3.0
djangorestframework>=3.12.4
zipseeker>=1.0.11
pandas>=1.3
sentry_sdk